feat: implement interactive selection for agents, models, and providers with display wrappers
