+++ crates/forge_main/Cargo.toml
@@ -51,6 +51,7 @@ open.workspace = true
 humantime.workspace = true
 num-format.workspace = true
 atty = "0.2"
+url.workspace = true
 
 [dev-dependencies]
 insta.workspace = true
