+++ crates/forge_main/src/config/mod.rs
@@ -1,6 +1,5 @@
 mod display;
 mod error;
 mod handler;
-mod interactive;
 
 pub use handler::handle_config_command;
